Review: restructuring + simplified (#563)
@gippy, thanks for the content! I went thru it and made some changes:
- It's docs, so we should have headings as titles, not questions, as
this is not FAQ, I know it was historically this way
- I updated the screenshots to be from the light version to be inlined
with the docs. That is a good point that I should put together some
instructions on this. I will put it into README.
- I also reshuffled the content to first introduce features and leave
"What should I monitor when scraping?" as the bonus.

Please, take a look and let me know what you think.


![obrazek](https://user-images.githubusercontent.com/594801/229557150-efd62f40-569d-4c83-92e3-96ea5ab1a363.png)