chore: update dependencies (#606)
Also fixes some links that were previously not reported by the older
docusaurus version. Unfortunately, there is a bug in the broken links
checker which results in false positives for some links, so had to
fallback to HTML links on some places. I guess its better than just
printing warnings on broken links, as that is easy to miss.

+ few random formatting and lint fixes to make the CI happy again