docs: add legal section (#932)
I'm adding a new top level section to Docs: `docs.apify.com/legal` This
section will house all of our public facing terms, policies and
agreements.

At this time, I don't need it to be linked from the Docs homepage or
from the top navigation bars, since the individual documents will be
linked directly from elsewhere. In the future, we should find a
reasonable place for a link to the whole section.

The individual terms and agreements will follow.