fix(docs): add babel plugin for styled components (#1001)
### 🐛 Bug

Some styling is completely messed up. This is most visible in the SDK
landing page for some reason. But also there is a padding missing in the
docs hompage.
The reason for this is that classes generated by styled-components are
not being added correctly.

### 🛠️ Fix

It seems that
[CSS-in-JS](https://docusaurus.io/docs/styling-layout#css-in-js) is not
supported by Docusaurus natively. I don't know why until now it has
somehow worked and started breaking just now...
In any case, by adding a babel plugin it seems that the code is now
transpiled correctly and the final build works with styled-components as
it should.