fix: fav icons (#1159)
Fixing missing icon in Google search (reported here
https://apify.slack.com/archives/C0L33UM7Z/p1724653919054129).

Added link for the `.ico` format. Also keep the link to the `.svg` with
`sizes: 'any'`.
Also updated the `.ico` to have margins (looks better in the Google
search with a circular border).

Seems like we are following the [Google
guidelines](https://developers.google.com/search/docs/appearance/favicon-in-search)
for icons in search. But the docs also says:
> A favicon isn't guaranteed in Google Search results, even if all
guidelines are met.