chore: remove patching
Previously there has been a patch to modify
front matter in Docusaurus v2, but with migration
to Docusaurus v3, this has been moved to a config
file and patching is not necessary anymore.
The patch itself has been removed from the codebase,
but I noticed there are leftovers in form of a dependency
and a postinstall script.