docs: Add documentation for Milvus/Zilliz database integration (#1203)
This integration is similar to Pinecone/Qdrant. I'm writing a tutorial
on how to use Apify with Milvus/Zilliz, and it would be good to
reference our documentation in the tutorial.

Just a note: Milvus is an open-source vector database, and Zilliz offers
a managed solution based on Milvus.

@TC-MO, could you please review the English? I haven't included any
screenshots, as I believe the description should suffice to get started.
Also, maintain and update screenshot is a bit painful.

Additionally, I haven't added this integration to the
[cards](https://docs.apify.com/platform/integrations#data-pipelines-etls-and-aillm-tools).
Do you think it should be included there?