feat: Add "Speeding up migrations" section to "State persistence" (#1365)
We've recently added support to speed up Actor migrations by rebooting
the run when it's migrating. This adds docs about that, with code
examples.

The support for this in the Apify SDK is not released yet, it'll go out
in the next SDK release, I'll only merge the docs then.

---------

Co-authored-by: Michał Olender <92638966+TC-MO@users.noreply.github.com>