fix: remove possible infinite loop in the changelog transformation (#1452)
In a recent `eslint` formatting PR (#1439), we changed the `void`
returns to `return 0`. This causes `unified` to loop indefinitely in
some cases (the visitor is expected to return `true / false` or **index
of the next sibling** to be visited).

This PR makes use of the exported `CONTINUE` symbol (alias for `true`)
and the more verbose return syntax.

Fixes the stalling documentation builds in Apify SDK for Python.