fix: remove the TypeScript course
The TypeScript course is an offtopic in the context of the Academy.
It is not a web scraping course, and it is not related to the Apify
platform per se. I see no reason why Apify should be teaching
TypeScript. I think we have enough more relevant docs to maintain.