feat: rename the JS course (#1551)
Rename the JS course from "Web scraping for beginners" to "Web scraping
basics for JavaScript devs", so that it is aligned with the Python
course and the design described in
https://github.com/apify/apify-docs/issues/1015

This change attempts to isolate the change to the course name only. The
name appears at many places which could be improved or which are
questionable, but this change is not intended to be a complete overhaul
of the course.