fix: re-order JSON and CSV in Python lessons (#1658)
When working on https://github.com/apify/apify-docs/pull/1584 I realized
it'd be better if the lesson started with JSON and continued with CSV,
not the other way.

In Python it doesn't matter and in JavaScript it's easier to start with
JSON, which is built-in, and only then move to CSV, which requires an
additional library. So for the sake of having both lessons aligned, I
want to change the order in the Python lesson, too.

So most of the diff is just the two sections reversed, and the two
exercises reversed. I made only a few additional changes to the wording.