feat: Improve description for MCP server tools configuration (#1773)
It should have only been a Section change (see context:
https://apify.slack.com/archives/C08N1LWC596/p1753972689045599).

But while I was at it, I took the liberty to make some changes to
improve clarity.

What I didn’t like about the current state:

- Tool configuration was mixed into the Dynamic .... section.
- Tool descriptions were essentially duplicated: once when describing
the tools themselves, and again when describing the parameter that
enabled the tools.

Link to the updated doc:
https://pr-1773.preview.docs.apify.com/platform/integrations/mcp#configure-tools-for-mcp-server

---------

Co-authored-by: Michał Olender <92638966+TC-MO@users.noreply.github.com>