feat: tidy up the tutorials navigation (#1873)
- I removed the PHP section containing a sole article. The article is
about using the Apify API from PHP, so I've put it inside the Apify API
section. I keep the slug untouched to avoid unnecessary redirects for
now. I redirect the root of the former PHP section directly to this
article, but not permanently, in case we would like to have the section
back again in the future.
- I changed the titles and H1s of the section and index pages so that
they're consistent. In titles, which are displayed on Google search
results, I wanted to have the "tutorials" word included, but in H1s,
which are displayed in the context of the navigation and docs UI, I
opted for a more brief heading.
- The articles in the Apify API section seem to be assorted, so I
removed `sidebar_position:` from their meta.
- "Apify scrapers" felt very vague, so I changed it to "Tutorials on
ready-made Apify scrapers", which is what the section is actually about.

I thought I could have a single section for Apify stuff in the
tutorials, where I'd move the
https://docs.apify.com/academy/running-a-web-server page, but the
"Tutorials on ready-made Apify scrapers" section is too self-contained
and the intro expects certain contents and ordering. I don't know
exactly yet what I'll do with it, but this PR feels like a good first
step, groundworks for further changes.