feat: serve markdown automatically based on `Accept` header (#1997)
Alters the nginx config so it checks the `Accept` header, if it contains
`text/plain` or `text/markdown`, we serve the markdown version directly.
Also added tests for this behavior to the existing PR check.

The previews are likely using the config from master, so there it
doesn't work yet (it should after we merge this).

Since the test workflow now effectively verifies the nginx config is
valid, I guess we can drop the `nginx.conf-test.yaml` workflow.