docs: Refactor of Build with AI section (#2044)
I refactored the page and added instructions on "how to start building
actors with AI".

Basically, the idea is to provide an environment for "vibe coders",
where they just copy and paste the instructions to AI, and the AI will
create an actor together with AGENTS.md.

Key decisions and comments:

- I ignored the apify create command and instead provided the LLM with
the actor structure with key details for actor.json, etc. It works
pretty well.
- I tested it, and I successfully built the actor several times. It also
created the correct AGENTS.md, and I was able to push the Actor to the
Console and run it.
- The next improvement regarding the AGENTS.md/instructions should be
done with the help of some kind of evals to systematically improve the
AGENTS.md/instructions rather than guessing what might work.
- I added links to install MCP through a button. There is a problem with
Auth (our MCP requires Auth even for docs). If we do not solve the
issue, I will just add a link to mcp.apify.com for now and add the
buttons later.

---------

Co-authored-by: Jiří Spilka <jiri.spilka@apify.com>
Co-authored-by: Michał Olender <92638966+TC-MO@users.noreply.github.com>